Fixed raycast not passing through doorways
Bullet was being blocked on the floor box collider.

Shooting script now checks for all collisions with raycast along path.
If collision is a Box collider, which is set to a trigger, and the collider is attached to a game object on the Layer:'GroundLayer', it is skipped as a collision.

It can now pass through the door ways, but does not pass through the mesh collider which will keep the collision effect on the floor.